+++ NEWS.md
@@ -1,3 +1,13 @@
+# semlbci 0.0.0.3
+
+* Can find the LBCI for free parameters and user-defined parameters.
+
+* Work for models with equality constraints.
+
+* Can specify parameters by lavaan model syntax.
+
+* Can use Wald CI to set the starting values. The optimization is usually faster.
+
 # semlbci 0.0.0.1
 
 * A first rough draft of the functions.